+++ app/lib/scorecard/models.dart
@@ -8,6 +8,7 @@ import 'package:json_annotation/json_annotation.dart';
 import 'package:meta/meta.dart';
 import 'package:pana/models.dart'
     show LicenseFile, PanaRuntimeInfo, Report, ReportSection;
+import 'package:pub_dev/scorecard/backend.dart';
 import 'package:pub_semver/pub_semver.dart';
 
 import '../dartdoc/models.dart';
@@ -95,6 +96,14 @@ class ScoreCard extends db.ExpandoModel<String> {
   @CompatibleStringListProperty()
   List<String> reportTypes = <String>[];
 
+  /// Compressed, json-encoded content of [PanaReport].
+  @db.BlobProperty()
+  List<int> panaReportJsonGz;
+
+  /// Compressed, json-encoded content of [DartdocReport].
+  @db.BlobProperty()
+  List<int> dartdocReportJsonGz;
+
   ScoreCard();
 
   ScoreCard.init({
@@ -123,6 +132,8 @@ class ScoreCard extends db.ExpandoModel<String> {
         derivedTags: derivedTags,
         flags: flags,
         reportTypes: reportTypes,
+        dartdocReport: DartdocReport.fromBytes(dartdocReportJsonGz),
+        panaReport: PanaReport.fromBytes(panaReportJsonGz),
       );
 
   Version get semanticRuntimeVersion => Version.parse(runtimeVersion);
@@ -197,12 +208,6 @@ class ScoreCardReport extends db.ExpandoModel<String> {
     reportJson = reportData.toJson();
   }
 
-  Map<String, dynamic> get reportJson {
-    if (reportJsonGz == null) return null;
-    return utf8JsonDecoder.convert(_gzipCodec.decode(reportJsonGz))
-        as Map<String, dynamic>;
-  }
-
   set reportJson(Map<String, dynamic> map) {
     if (map == null) {
       reportJsonGz = null;
@@ -214,9 +219,9 @@ class ScoreCardReport extends db.ExpandoModel<String> {
   ReportData get reportData {
     switch (reportType) {
       case ReportType.pana:
-        return PanaReport.fromJson(reportJson);
+        return PanaReport.fromBytes(reportJsonGz);
       case ReportType.dartdoc:
-        return DartdocReport.fromJson(reportJson);
+        return DartdocReport.fromBytes(reportJsonGz);
     }
     throw Exception('Unknown report type: $reportType');
   }
@@ -269,6 +274,9 @@ class ScoreCardData extends Object with FlagMixin {
   /// The report types that are already done for the ScoreCard.
   final List<String> reportTypes;
 
+  final DartdocReport dartdocReport;
+  final PanaReport panaReport;
+
   ScoreCardData({
     this.packageName,
     this.packageVersion,
@@ -282,6 +290,8 @@ class ScoreCardData extends Object with FlagMixin {
     this.derivedTags,
     this.flags,
     this.reportTypes,
+    this.dartdocReport,
+    this.panaReport,
   });
 
   factory ScoreCardData.fromJson(Map<String, dynamic> json) =>
@@ -350,6 +360,13 @@ class PanaReport implements ReportData {
   factory PanaReport.fromJson(Map<String, dynamic> json) =>
       _$PanaReportFromJson(json);
 
+  static PanaReport fromBytes(List<int> bytes) {
+    if (bytes == null) return null;
+    final map = utf8JsonDecoder.convert(_gzipCodec.decode(bytes))
+        as Map<String, dynamic>;
+    return PanaReport.fromJson(map);
+  }
+
   @override
   Map<String, dynamic> toJson() => _$PanaReportToJson(this);
 }
@@ -376,6 +393,13 @@ class DartdocReport implements ReportData {
 
   factory DartdocReport.fromJson(Map<String, dynamic> json) =>
       _$DartdocReportFromJson(json);
+
+  static DartdocReport fromBytes(List<int> bytes) {
+    if (bytes == null) return null;
+    final map = utf8JsonDecoder.convert(_gzipCodec.decode(bytes))
+        as Map<String, dynamic>;
+    return DartdocReport.fromJson(map);
+  }
 
   @override
   Map<String, dynamic> toJson() => _$DartdocReportToJson(this);